# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: OpenLedger API
  version: 0.1.0
  description: >
    GnuCash-derived double-entry accounting API. This spec is the source of
    truth for request/response shapes; the web frontend consumes a generated
    TypeScript client. Money is always exact rational num/denom pairs, never a
    JSON float.
servers:
  - url: http://openledger.localhost
# Authentication is handled at the proxy layer (Traefik + Authelia forward-auth).
# Authelia sets Remote-User/Remote-Email headers that the API trusts; there are
# no /auth/* endpoints and no JWT tokens in the API itself.
# Users log in and out via http://auth.openledger.localhost.
paths:
...
  /api/v1/books/{id}/invoices:
    get:
      summary: List invoices or bills for a book
      operationId: listInvoices
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [invoice, bill]
            default: invoice
      responses:
        "200":
          description: Invoice list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceList"
    post:
      summary: Create a draft invoice or bill
      operationId: createInvoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewInvoice"
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /api/v1/invoices/{id}:
    get:
      summary: Get an invoice with its entries
      operationId: getInvoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Invoice not found
    patch:
      summary: Update a draft invoice
      operationId: updateInvoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewInvoice"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Invoice not found
        "422":
          description: Invoice already posted
    delete:
      summary: Delete a draft invoice
      operationId: deleteInvoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Invoice not found
        "422":
          description: Invoice already posted
  /api/v1/invoices/{id}/pay:
    post:
      summary: Record a payment against a posted invoice
      operationId: payInvoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayInvoiceRequest"
      responses:
        "200":
          description: Paid invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Invoice not found
        "422":
          description: Not posted or already paid
  /api/v1/invoices/{id}/post:
    post:
      summary: Post an invoice to the ledger
      operationId: postInvoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInvoiceRequest"
      responses:
        "200":
          description: Posted invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Invoice not found
        "422":
          description: Already posted or no entries
  /api/v1/invoices/{id}/entries:
    get:
      summary: List entries for an invoice
      operationId: listEntries
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Entry list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntryList"
    post:
      summary: Add a line item to a draft invoice
      operationId: addEntry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewEntry"
      responses:
        "201":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "422":
          description: Invoice already posted
  /api/v1/entries/{id}:
    patch:
      summary: Update a line item
      operationId: updateEntry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewEntry"
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "404":
          description: Entry not found
        "422":
          description: Invoice already posted
    delete:
      summary: Delete a line item
      operationId: deleteEntry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Entry not found
        "422":
          description: Invoice already posted
components:
  schemas:
...
    NewInvoice:
      type: object
      required: [type, ownerGuid, currencyGuid]
      properties:
        id:
          type: string
          description: Display number (e.g. INV-0001)
        type:
          type: string
          enum: [invoice, bill]
        ownerGuid:
          type: string
          description: Customer GUID (invoice) or Vendor GUID (bill)
        dateOpened:
          type: string
          format: date
        notes:
          type: string
        active:
          type: boolean
          default: true
        currencyGuid:
          type: string
        termsGuid:
          type: string
    Invoice:
      allOf:
        - $ref: "#/components/schemas/NewInvoice"
        - type: object
          properties:
            guid:
              type: string
            bookGuid:
              type: string
            datePosted:
              type: string
              format: date
              nullable: true
            dateDue:
              type: string
              format: date
              nullable: true
            postTxnGuid:
              type: string
            postAccGuid:
              type: string
            paidAt:
              type: string
              format: date-time
              nullable: true
            paidTxnGuid:
              type: string
            createdAt:
              type: string
              format: date-time
            entries:
              type: array
              items:
                $ref: "#/components/schemas/Entry"
    InvoiceList:
      type: object
      properties:
        bookGuid:
          type: string
        type:
          type: string
        invoices:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
    PostInvoiceRequest:
      type: object
      required: [postAccGuid]
      properties:
        postDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        postAccGuid:
          type: string
          description: A/R account GUID (invoice) or A/P account GUID (bill)
    NewEntry:
      type: object
      required: [accountGuid, price]
      properties:
        date:
          type: string
          format: date
        description:
          type: string
        action:
          type: string
        notes:
          type: string
        quantity:
          $ref: "#/components/schemas/GncNumeric"
        accountGuid:
          type: string
        price:
          $ref: "#/components/schemas/GncNumeric"
        taxable:
          type: boolean
          default: false
        taxTableGuid:
          type: string
          description: tax table applied when taxable; empty for no tax
    Entry:
      allOf:
        - $ref: "#/components/schemas/NewEntry"
        - type: object
          properties:
            guid:
              type: string
            invoiceGuid:
              type: string
            lineTotal:
              $ref: "#/components/schemas/GncNumeric"
            createdAt:
              type: string
              format: date-time
    EntryList:
      type: object
      properties:
        invoiceGuid:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/Entry"
    PayInvoiceRequest:
      type: object
      required: [paymentAccGuid]
      properties:
        paymentAccGuid:
          type: string
          description: Cash or bank account GUID to debit (customer invoice) or credit (vendor bill)
        paymentDate:
          type: string
          format: date
...